probe class is now in acq module
